Rename: 부스 페이지 내부에서 사용하는 컴포넌트 ui 파일 이름 수정, 부스에서 사용되는 페이지들 디렉토리로 관리 시작
