+++ src/components/ui/booth-category.tsx
